Move ancillary classes into a separate file
